Move locale init to index.ts before app loading
